Bluetooth: Mesh: Move scene entries to flash
Redesign of the scene entry feature.

Instead of keeping a separate linked list of scene entry objects in the
model contexts, we'll trade RAM for CPU, and perform a lookup in
ROM-allocated scene entries for each model in the composition data when
storing and restoring.

This removes all RAM usage for scene entries, saving 16 bytes per model,
for all models with scene data. As there'll only be one entry for each
model type, we'll have to search through at most 10 entries per model.

Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>