applications: asset_tracker_v2: Do not reboot on sensor init failure.
Failing to initialize external sensors should not be considered a
irrecoverable error that should be rebooted on.

This patch removes functionality that returns error upon sensor
initialization errors and adds callbacks to the ext_sensors library to
signal the sensor module which sensor that failed initialization.
Future improvement is to send a message to cloud in case issues with
external sensors occurs.

Closes CIA-324

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>