kconfig: Provide minimal TF-M configuration
To make TF-M a more attractive alternative to the home-brewn SPM
we need to provide a minimal variant which shares the feature set
and which has the same memory footprint.

One major difference between SPM and TF-M is that the latter is
built as an external project, not as a 'child image'. This difference
greatly affects how the images are configured.

While SPM is configured by passing options to a child image, TF-M
is configured through the 'zephyr module' integration with the
'trusted-firmware-m' module.

As a result of this, it is not trivial to provide an easy-to-include
minimal configuration for TF-M. This commits adds several defconfig
files which will create a minimal TF-M configuration. It also
includes patches needed for not reserving memory regions that would
be unused in a minimal build of TF-M.

Ref: NCSDK-7561

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>