net: lib: nrf_cloud: failed update handling
If nrf_cloud_pgps_request_all() fails in
nrf_cloud_pgps_notify_prediction(), a flag is left in the wrong state,
and no further attempt is made to download predictions until the device
resets.

This clears the flag, which allows the next call to this function to
retry.

Jira: NCSDK-9922

Signed-off-by: Pete Skeggs <peter.skeggs@nordicsemi.no>