zigbee: Provide a special reset type for NCP SoC
In case of reset that is triggered by NCP host (ZBOSS), the NCP SoC
shall send Reset Response frame instead of Reset Indication.
The default logic is based on the reset type, so all soft resets are
treated as a result of NCP reset command, which is not true, since the
soft reset may be triggered by other tasks or a core dump.

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>