spm: Fix function call for FP register storing
The function used an old signature for the context storing function.
The new function has just one argument, and by chance, the first
parameter passed with the wrong signature was a buffer with length
of 1 word less than the struct in the new signature, which meant that
the function was clobbering the first word after the buffer.

Ref: NCSIDB-486

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>