Bluetooth: Mesh: Refactor Light XYL and HSL Server models
When a node implements Light HSL Server and Light xyL Server located on
the same element, the Light xyL state and the Light HSL state are bound
indirectly via the Light Lightness Actual state.

The user has to add Lightness Server model to the composition data
separately and pass a pointer to HSL and XYL Server init macros.

Signed-off-by: Michał Narajowski <michal.narajowski@codecoup.pl>