lib: agps: add own API for the A-GPS library
Removed A-GPS library functions from the GPS driver header, added an
own header file for the A-GPS library and renamed functions in the API
accordingly. Removed dependency to the GPS driver from the library API,
instead a GNSS API structure is used for providing the requested A-GPS
data to the library.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>