scripts: get_pulls_in_range: handle commits merged without PR
This upmerge is the second time that I'm aware of upstream Zephyr
introducing commits without a pull request. This situation currently
causes the script to fail. The last time it happened, I asked on the
upstream slack and was told it wouldn't happen again. That's clearly
false, so just handle the situation with a special marker value that
says "this commit didn't come from a PR".

Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>