applications: asset_tracker_v2: Add parameters to module events
Add number of satellites tracked and GPS search time for fixes and
timeouts to events sent out by the module.

Functionality has been added to the module itself to keep track of
these new parameters. Ideally we should get this information
directly from the driver, but due to a lot of pending changes
to how GNSS is handled in the application this functionality is
temporarily handled in the GPS module.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>