doc: update Doxygen requirements
Doxygen version >= 1.8.18 is now required. 1.8.17 version, the one
shipped with latest Ubuntu LTS has some issues when handling missing
return docs for void functions. This causes warnings on the nrfx docset,
so that version has to be skipped. The latest version (1.9.2) has been
chosen as Doxygen does not seem to provide a URL to download 1.8.18
built for Ubuntu 20.04.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>