application: pelion client: use new configuration and pm conf scheme
This commit reworks Pelion Client to support the APPLICATION_CONFIG_DIR.

The APPLICATION_CONFIG_DIR has the same structure as Zephyr normal
Kconfig and overlay files as they are place in the sample folder, but
allows the application to redirect this structure to a custom location.

For nRF Connect SDK the directory `<app>/configuration/<board>` is a
common configuration folder for more complex applications.

This removes the need for several lines of boilerplate code in the
application's CMakeLists.txt file and at the same time allows us to
unify an alternative configuration structure in a common way.

It also uses the extended pm_static naming scheme where a static pm
configuration file can be named according to the board and revision for
which it should be used.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>