samples: Fix samples with missing integration_platforms
Add a list of integration_platfoms to samples that are still
missing it. This is an important fix for proper and complying with
the workflow operation of the sdk-nrf CI. Samples without this list
leave a lot of noise in the result reports as Twister sets the
scope for them to all Zephyr's platforms (~400) and then filters
the vast majority of them resulting in a huge amount of skips in
the results. It is also crucial to assure that the listed
platforms will not get skipped.

Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>