net: lib: azure_iot_hub: Increase default topic element max size
Messages sent from the Azure IoT Explorer
to a device now comes with larger property bags than before,
resulting in error messages.
Increase the max length for topic elements to match the required
length.
The stack also needs to be bumped to process the larger elements.

Tests are also updated.

Closes CIA-416

Signed-off-by: Jan Tore Guggedal <jantore.guggedal@nordicsemi.no>