net: lib: azure_iot_hub: add content-type and encoding to messages
This adds the content-type and encoding to the messages
sent by the device.

This is needed to enable message based routing on the
cloud side.

Routing based on property bags would be ideal for the
Asset Tracker v2 but this is not supported on Azure IoT.

See https://azure.microsoft.com/es-es/blog/iot-hub-message-routing-now-with-routing-on-message-body/

Signed-off-by: Markus Tacker <Markus.Tacker@NordicSemi.no>