lib: at_monitor: subscriber list alignment fix
struct at_monitor_entry is always aligned on word boundary. When
_at_monitor_entry_list_start is placed at unaligned address, it will
point to aligning bytes and AT monitor would not find entries to send
notifications to. This depends on the size of previous section, and so
will occur randomly on some builds. Add ALINGN(4) directive to ensure
that the _at_monitor_entry_list_start will point to the beginning of the
list.

Signed-off-by: Valentin Plotkin <valentin.plotkin@trisense.no>