caf: power_manager: Depend configuration on PM_POLICY_APP
This commit adds direct dependency of power manager on PM_POLICY_APP.
PM_POLICY_APP has to be enabled for power manager to link properly.
We cannot select PM_POLICY_APP symbol here as it is choice symbol.
This dependency gives the user easy to read hint that this configuration
is required.

Signed-off-by: Radoslaw Koppel <radoslaw.koppel@nordicsemi.no>