kconfig: partition_manager: set size of mcuboot secondary automatically
Fix bug in partition manager where 'share_size' could not be used for
sharing size between a partition in a region with non-complex placement
strategy and a partition in a region with complex placement strategy.

Also leverage this functionality to avoid having to manually set the
size of the mcuboot secondary slot when it is stored in external flash.

Ref: NCSDK-10375

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>