cmake: use CACHE STRING FORCE for add_overlay function
The use of
> set(${image}_${overlay_type} "<val>" CACHE INTERNAL "")
has been modified to use
> set({image}_${overlay_type} "<val>" CACHE STRING "" FORCE)
instead.

According to CMake documentation, the INTERNAL flag implies FORCE.
However, as NCDIDB-608 reveals, there seems to be an edge case when the
user initializes the cache variable and the build system later updates
it using INTERNAL.
This behavior appears to be scope dependent as it works as expected in
most cases.

Specifically setting the cache variable to type `STRING` and specify
`FORCE` works as expected and updates the CMake cache correctly in all
cases.

This has been filed as an official CMake bug here:
https://gitlab.kitware.com/cmake/cmake/-/issues/22734

Fixes: NCSIDB-608

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>